Add files via upload
"This portfolio is a showcase of my journey as a frontend web developer. It features various projects that highlight my skills in HTML, CSS, and JavaScript, as well as my commitment to learning and improving in the web development field. The portfolio includes a range of interactive and responsive web designs, demonstrating my ability to create clean and user-friendly interfaces. I’ve built this portfolio to not only showcase my work but also as a way to push my boundaries and experiment with new technologies. The goal is to continuously enhance my skills and provide high-quality web experiences for users."